+++ src/com/maddyhome/idea/vim/VimTypedActionHandler.java
@@ -21,10 +21,11 @@ package com.maddyhome.idea.vim;
 import com.intellij.codeInsight.lookup.Lookup;
 import com.intellij.codeInsight.lookup.LookupManager;
 import com.intellij.openapi.actionSystem.DataContext;
-import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.diagnostic.Logger;
+import com.intellij.openapi.editor.actionSystem.ActionPlan;
 import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.editor.actionSystem.TypedActionHandler;
+import com.intellij.openapi.editor.actionSystem.TypedActionHandlerEx;
 import com.maddyhome.idea.vim.helper.EditorDataContext;
 import org.jetbrains.annotations.NotNull;
 
@@ -35,7 +36,7 @@ import javax.swing.*;
  *
  * IDE shortcut keys used by Vim commands are handled by {@link com.maddyhome.idea.vim.action.VimShortcutKeyAction}.
  */
-public class VimTypedActionHandler implements TypedActionHandler {
+public class VimTypedActionHandler implements TypedActionHandlerEx {
   private static final Logger logger = Logger.getInstance(VimTypedActionHandler.class.getName());
 
   private final TypedActionHandler origHandler;
@@ -48,20 +49,27 @@ public class VimTypedActionHandler implements TypedActionHandler {
   }
 
   @Override
+  public void beforeExecute(@NotNull Editor editor, char charTyped, @NotNull DataContext context, @NotNull ActionPlan plan) {
+    if (isEnabled(editor)) {
+      handler.beforeHandleKey(editor, KeyStroke.getKeyStroke(charTyped), context, plan);
+    }
+    else {
+      TypedActionHandler originalHandler = KeyHandler.getInstance().getOriginalHandler();
+      if (originalHandler instanceof TypedActionHandlerEx) {
+        ((TypedActionHandlerEx)originalHandler).beforeExecute(editor, charTyped, context, plan);
+      }
+    }
+  }
+
+  @Override
   public void execute(@NotNull final Editor editor, final char charTyped, @NotNull final DataContext context) {
     if (isEnabled(editor)) {
-      // Run key handler outside of the key typed command for creating our own undoable commands
-      ApplicationManager.getApplication().invokeLater(new Runnable() {
-        @Override
-        public void run() {
-          try {
-            handler.handleKey(editor, KeyStroke.getKeyStroke(charTyped), new EditorDataContext(editor));
-          }
-          catch (Throwable e) {
-            logger.error(e);
-          }
-        }
-      });
+      try {
+        handler.handleKey(editor, KeyStroke.getKeyStroke(charTyped), new EditorDataContext(editor));
+      }
+      catch (Throwable e) {
+        logger.error(e);
+      }
     }
     else {
       origHandler.execute(editor, charTyped, context);