+++ src/com/maddyhome/idea/vim/ex/handler/ActionHandler.java
@@ -26,6 +26,7 @@ import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.editor.Editor;
 import com.maddyhome.idea.vim.KeyHandler;
 import com.maddyhome.idea.vim.VimPlugin;
+import com.maddyhome.idea.vim.command.CommandState;
 import com.maddyhome.idea.vim.ex.CommandHandler;
 import com.maddyhome.idea.vim.ex.ExCommand;
 import com.maddyhome.idea.vim.ex.ExException;
@@ -37,7 +38,7 @@ import org.jetbrains.annotations.NotNull;
  */
 public class ActionHandler extends CommandHandler {
   public ActionHandler() {
-    super("action", "", RANGE_FORBIDDEN | DONT_REOPEN);
+    super("action", "", RANGE_OPTIONAL | DONT_REOPEN);
   }
 
   public boolean execute(@NotNull Editor editor, @NotNull final DataContext context,
@@ -50,26 +51,31 @@ public class ActionHandler extends CommandHandler {
     }
     final Application application = ApplicationManager.getApplication();
     if (application.isUnitTestMode()) {
-      executeAction(action, context, actionName);
+      executeAction(editor, cmd, action, context, actionName);
     }
     else {
-      UiHelper.runAfterGotFocus(new Runnable() {
-        @Override
-        public void run() {
-          executeAction(action, context, actionName);
-        }
-      });
+      UiHelper.runAfterGotFocus(() -> executeAction(editor, cmd, action, context, actionName));
     }
     return true;
   }
 
-  private void executeAction(@NotNull AnAction action, @NotNull DataContext context, @NotNull String actionName) {
+  private void executeAction(@NotNull Editor editor, @NotNull ExCommand cmd, @NotNull AnAction action,
+                             @NotNull DataContext context, @NotNull String actionName) {
+    final boolean visualAction = cmd.getRanges().size() > 0;
+    if (visualAction) {
+      VimPlugin.getMotion().selectPreviousVisualMode(editor);
+    }
     try {
       KeyHandler.executeAction(action, context);
     }
     catch (RuntimeException e) {
-      // TODO: Find out if any runtime exceptions may happen here
       assert false : "Error while executing :action " + actionName + " (" + action + "): " + e;
+    }
+    finally {
+      if (visualAction) {
+        // Exit visual mode selected above, but do it without resetting the selected text
+        CommandState.getInstance(editor).popState();
+      }
     }
   }
 }