+++ src/com/maddyhome/idea/vim/group/MotionGroup.java
@@ -21,15 +21,15 @@ import com.intellij.openapi.actionSystem.DataContext;
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.editor.*;
 import com.intellij.openapi.editor.event.*;
+import com.intellij.openapi.editor.ex.DocumentEx;
 import com.intellij.openapi.editor.ex.util.EditorUtil;
-import com.intellij.openapi.fileEditor.FileEditor;
-import com.intellij.openapi.fileEditor.FileEditorManagerAdapter;
-import com.intellij.openapi.fileEditor.FileEditorManagerEvent;
-import com.intellij.openapi.fileEditor.TextEditor;
+import com.intellij.openapi.fileEditor.*;
 import com.intellij.openapi.fileEditor.impl.EditorTabbedContainer;
 import com.intellij.openapi.fileEditor.impl.EditorWindow;
 import com.intellij.openapi.vfs.LocalFileSystem;
 import com.intellij.openapi.vfs.VirtualFile;
+import com.intellij.openapi.vfs.VirtualFileManager;
+import com.intellij.openapi.vfs.VirtualFileSystem;
 import com.maddyhome.idea.vim.EventFacade;
 import com.maddyhome.idea.vim.KeyHandler;
 import com.maddyhome.idea.vim.VimPlugin;
@@ -399,7 +399,7 @@ public class MotionGroup {
       }
       final int line = mark.getLogicalLine();
       if (!vf.getPath().equals(mark.getFilename())) {
-        final Editor selectedEditor = selectEditor(editor, vf);
+        final Editor selectedEditor = selectEditor(editor, mark);
         if (selectedEditor != null) {
           moveCaret(selectedEditor, moveCaretToLineStartSkipLeading(selectedEditor, line));
         }
@@ -434,7 +434,7 @@ public class MotionGroup {
       }
       final LogicalPosition lp = new LogicalPosition(mark.getLogicalLine(), mark.getCol());
       if (!vf.getPath().equals(mark.getFilename())) {
-        final Editor selectedEditor = selectEditor(editor, vf);
+        final Editor selectedEditor = selectEditor(editor, mark);
         if (selectedEditor != null) {
           moveCaret(selectedEditor, selectedEditor.logicalPositionToOffset(lp));
         }
@@ -495,6 +495,26 @@ public class MotionGroup {
     else {
       return -1;
     }
+  }
+
+  @Nullable
+  private Editor selectEditor(@NotNull Editor editor, @NotNull Mark mark) {
+    final VirtualFile virtualFile = markToVirtualFile(mark);
+    if (virtualFile != null) {
+      return selectEditor(editor, virtualFile);
+    } else {
+      return null;
+    }
+  }
+
+  @Nullable
+  private VirtualFile markToVirtualFile(@NotNull Mark mark) {
+    String protocol = mark.getProtocol();
+    VirtualFileSystem fileSystem = VirtualFileManager.getInstance().getFileSystem(protocol);
+    if (mark.getFilename() == null) {
+      return null;
+    }
+    return fileSystem.findFileByPath(mark.getFilename());
   }
 
   @Nullable
@@ -1649,8 +1669,8 @@ public class MotionGroup {
       final BoundStringOption opt = (BoundStringOption)Options.getInstance().getOption("selection");
       int lineEnd = EditorHelper.getLineEndForOffset(editor, end);
       final int adj = opt.getValue().equals("exclusive") || end == lineEnd ? 0 : 1;
-      end = Math.min(EditorHelper.getFileSize(editor), end + adj);
-      editor.getSelectionModel().setSelection(start, end);
+      final int adjEnd = Math.min(EditorHelper.getFileSize(editor), end + adj);
+      editor.getSelectionModel().setSelection(start, adjEnd);
     }
     else if (subMode == CommandState.SubMode.VISUAL_LINE) {
       if (start > end) {
@@ -1728,15 +1748,16 @@ public class MotionGroup {
     private boolean myMakingChanges = false;
 
     public void selectionChanged(@NotNull SelectionEvent selectionEvent) {
-      if (myMakingChanges) {
+      final Editor editor = selectionEvent.getEditor();
+      final Document document = editor.getDocument();
+      if (myMakingChanges || (document instanceof DocumentEx && ((DocumentEx)document).isInEventsHandling())) {
         return;
       }
 
       myMakingChanges = true;
       try {
-        final Editor editor = selectionEvent.getEditor();
         final com.intellij.openapi.util.TextRange newRange = selectionEvent.getNewRange();
-        for (Editor e : EditorFactory.getInstance().getEditors(editor.getDocument())) {
+        for (Editor e : EditorFactory.getInstance().getEditors(document)) {
           if (!e.equals(editor)) {
             e.getSelectionModel().setSelection(newRange.getStartOffset(), newRange.getEndOffset());
           }
@@ -1807,6 +1828,14 @@ public class MotionGroup {
     @NotNull private CommandState.SubMode mode = CommandState.SubMode.NONE;
     private int startOff;
     private int endOff;
+  }
+
+  public int getLastFTCmd() {
+    return lastFTCmd;
+  }
+
+  public char getLastFTChar() {
+    return lastFTChar;
   }
 
   private int lastFTCmd = 0;