+++ src/com/maddyhome/idea/vim/helper/UiHelper.kt
@@ -1,0 +1,49 @@
+/*
+ * IdeaVim - Vim emulator for IDEs based on the IntelliJ platform
+ * Copyright (C) 2003-2016 The IdeaVim authors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+@file:JvmName("UiHelper")
+
+package com.maddyhome.idea.vim.helper
+
+import com.intellij.openapi.application.ModalityState
+import com.intellij.openapi.editor.colors.EditorColorsManager
+import com.intellij.openapi.wm.IdeFocusManager
+import java.awt.Font
+import javax.swing.JComponent
+
+/**
+ * Get focus reliably.
+ */
+fun requestFocus(component: JComponent) {
+  IdeFocusManager.findInstance().requestFocus(component, true)
+}
+
+/**
+ * Run code after getting focus on request.
+ *
+ * @see [requestFocus]
+ */
+fun runAfterGotFocus(runnable: Runnable) {
+  IdeFocusManager.findInstance().doWhenFocusSettlesDown(runnable, ModalityState.defaultModalityState())
+}
+
+val editorFont: Font
+  get() {
+    val scheme = EditorColorsManager.getInstance().globalScheme
+    return Font(scheme.editorFontName, Font.PLAIN, scheme.editorFontSize)
+  }