+++ src/com/maddyhome/idea/vim/option/Options.java
@@ -473,6 +473,7 @@ public class Options {
     addOption(new ToggleOption(TIMEOUT, "to", true));
     addOption(new ListOption(VIMINFO, "vi", new String[]{"'100", "<50", "s10", "h"}, null));
     addOption(new StringOption(GVIM_PATH, "gvpath", ""));
+    addOption(new KeywordOption("iskeyword", "isk", new String[]{"@", "48-57", "_"}));
 
     registerExtensionOptions();
   }
@@ -506,8 +507,8 @@ public class Options {
     abbrevs.put(option.getAbbreviation(), option);
   }
 
-  @NotNull private final HashMap<String, Option> options = new HashMap<String, Option>();
-  @NotNull private final HashMap<String, Option> abbrevs = new HashMap<String, Option>();
+  @NotNull private final HashMap<String, Option> options = new LinkedHashMap<>();
+  @NotNull private final HashMap<String, Option> abbrevs = new LinkedHashMap<>();
 
   private static Options ourInstance;